UPD: decompose APS into Naive, APS and RAPS + new abtract methods for classification
