+++ mapie/conformity_scores/__init__.py
@@ -3,7 +3,7 @@ from .classification import BaseClassificationScore
 from .bounds import (
     AbsoluteConformityScore, GammaConformityScore, ResidualNormalisedScore
 )
-from .sets import APS, LAC, TopK
+from .sets import APS, LAC, Naive, RAPS, TopK
 
 
 __all__ = [
@@ -12,7 +12,9 @@ __all__ = [
     "AbsoluteConformityScore",
     "GammaConformityScore",
     "ResidualNormalisedScore",
+    "Naive",
     "LAC",
     "APS",
+    "RAPS",
     "TopK"
 ]
