+++ mapie/conformity_scores/sets/__init__.py
@@ -1,10 +1,14 @@
+from .naive import Naive
 from .lac import LAC
 from .aps import APS
+from .raps import RAPS
 from .topk import TopK
 
 
 __all__ = [
+    "Naive",
     "LAC",
     "APS",
+    "RAPS",
     "TopK",
 ]
