+++ mapie/conformity_scores/sets/utils.py
@@ -1,7 +1,6 @@
-from typing import Any, Optional, Tuple, Union, cast
+from typing import Optional, Tuple, Union, cast
 import numpy as np
 from sklearn.calibration import label_binarize
-from sklearn.dummy import check_random_state
 
 from mapie._typing import ArrayLike, NDArray
 from mapie._machine_precision import EPSILON
@@ -203,199 +202,3 @@ def get_last_index_included(
             ), axis=1
         )
     return y_pred_index_last[:, np.newaxis, :]
-
-
-def get_last_included_proba(
-    y_pred_proba: NDArray,
-    thresholds: NDArray,
-    include_last_label: Union[bool, str, None],
-    method: str,
-    lambda_: Union[NDArray, float, None],
-    k_star: Union[NDArray, Any]
-) -> Tuple[NDArray, NDArray, NDArray]:
-    """
-    Function that returns the smallest score
-    among those which are included in the prediciton set.
-
-    Parameters
-    ----------
-    y_pred_proba: NDArray of shape (n_samples, n_classes)
-        Predictions of the model.
-
-    thresholds: NDArray of shape (n_alphas, )
-        Quantiles that have been computed from the conformity scores.
-
-    include_last_label: Union[bool, str, None]
-        Whether to include or not the label whose score exceeds the threshold.
-
-    lambda_: Union[NDArray, float, None] of shape (n_alphas)
-        Values of lambda for the regularization.
-
-    k_star: Union[NDArray, Any]
-        Values of k for the regularization.
-
-    Returns
-    -------
-    Tuple[ArrayLike, ArrayLike, ArrayLike]
-        Arrays of shape (n_samples, n_classes, n_alphas),
-        (n_samples, 1, n_alphas) and (n_samples, 1, n_alphas).
-        They are respectively the cumsumed scores in the original
-        order which can be different according to the value of alpha
-        with the RAPS method, the index of the last included score
-        and the value of the last included score.
-    """
-    index_sorted = np.flip(
-        np.argsort(y_pred_proba, axis=1), axis=1
-    )
-    # sort probabilities by decreasing order
-    y_pred_proba_sorted = np.take_along_axis(
-        y_pred_proba, index_sorted, axis=1
-    )
-    # get sorted cumulated score
-    y_pred_proba_sorted_cumsum = np.cumsum(
-        y_pred_proba_sorted, axis=1
-    )
-
-    if method == "raps":
-        y_pred_proba_sorted_cumsum += lambda_ * np.maximum(
-            0,
-            np.cumsum(
-                np.ones(y_pred_proba_sorted_cumsum.shape), axis=1
-            ) - k_star
-        )
-    # get cumulated score at their original position
-    y_pred_proba_cumsum = np.take_along_axis(
-        y_pred_proba_sorted_cumsum,
-        np.argsort(index_sorted, axis=1),
-        axis=1
-    )
-    # get index of the last included label
-    y_pred_index_last = get_last_index_included(
-        y_pred_proba_cumsum,
-        thresholds,
-        include_last_label
-    )
-    # get the probability of the last included label
-    y_pred_proba_last = np.take_along_axis(
-        y_pred_proba,
-        y_pred_index_last,
-        axis=1
-    )
-
-    zeros_scores_proba_last = (y_pred_proba_last <= EPSILON)
-
-    # If the last included proba is zero, change it to the
-    # smallest non-zero value to avoid inluding them in the
-    # prediction sets.
-    if np.sum(zeros_scores_proba_last) > 0:
-        y_pred_proba_last[zeros_scores_proba_last] = np.expand_dims(
-            np.min(
-                np.ma.masked_less(
-                    y_pred_proba,
-                    EPSILON
-                ).filled(fill_value=np.inf),
-                axis=1
-            ), axis=1
-        )[zeros_scores_proba_last]
-
-    return y_pred_proba_cumsum, y_pred_index_last, y_pred_proba_last
-
-
-def add_random_tie_breaking(
-    prediction_sets: NDArray,
-    y_pred_index_last: NDArray,
-    y_pred_proba_cumsum: NDArray,
-    y_pred_proba_last: NDArray,
-    threshold: NDArray,
-    method: str,
-    random_state: Optional[Union[int, np.random.RandomState]] = None,
-    lambda_star: Optional[Union[NDArray, float]] = None,
-    k_star: Optional[Union[NDArray, None]] = None
-) -> NDArray:
-    """
-    Randomly remove last label from prediction set based on the
-    comparison between a random number and the difference between
-    cumulated score of the last included label and the quantile.
-
-    Parameters
-    ----------
-    prediction_sets: NDArray of shape
-        (n_samples, n_classes, n_threshold)
-        Prediction set for each observation and each alpha.
-
-    y_pred_index_last: NDArray of shape (n_samples, threshold)
-        Index of the last included label.
-
-    y_pred_proba_cumsum: NDArray of shape (n_samples, n_classes)
-        Cumsumed probability of the model in the original order.
-
-    y_pred_proba_last: NDArray of shape (n_samples, 1, threshold)
-        Last included probability.
-
-    threshold: NDArray of shape (n_alpha,) or shape (n_samples_train,)
-        Threshold to compare with y_proba_last_cumsum, can be either:
-
-        - the quantiles associated with alpha values when ``cv`` == "prefit",
-            ``cv`` == "split" or ``agg_scores`` is "mean"
-
-        - the conformity score from training samples otherwise
-            (i.e., when ``cv`` is CV splitter and ``agg_scores`` is "crossval")
-
-    method: str
-        Method that determines how to remove last label in the prediction set.
-
-        - if "cumulated_score" or "aps", compute V parameter from Romano+(2020)
-
-        - else compute V parameter from Angelopoulos+(2020)
-
-    lambda_star: Union[NDArray, float, None] of shape (n_alpha):
-        Optimal value of the regulizer lambda.
-
-    k_star: Union[NDArray, None] of shape (n_alpha):
-        Optimal value of the regulizer k.
-
-    Returns
-    -------
-    NDArray of shape (n_samples, n_classes, n_alpha)
-        Updated version of prediction_sets with randomly removed labels.
-    """
-    # get cumsumed probabilities up to last retained label
-    y_proba_last_cumsumed = np.squeeze(
-        np.take_along_axis(
-            y_pred_proba_cumsum,
-            y_pred_index_last,
-            axis=1
-        ), axis=1
-    )
-
-    if method in ["cumulated_score", "aps"]:
-        # compute V parameter from Romano+(2020)
-        vs = (
-            (y_proba_last_cumsumed - threshold.reshape(1, -1)) /
-            y_pred_proba_last[:, 0, :]
-        )
-    else:
-        # compute V parameter from Angelopoulos+(2020)
-        L = np.sum(prediction_sets, axis=1)
-        vs = (
-            (y_proba_last_cumsumed - threshold.reshape(1, -1)) /
-            (
-                y_pred_proba_last[:, 0, :] -
-                lambda_star * np.maximum(0, L - k_star) +
-                lambda_star * (L > k_star)
-            )
-        )
-
-    # get random numbers for each observation and alpha value
-    random_state = check_random_state(random_state)
-    random_state = cast(np.random.RandomState, random_state)
-    us = random_state.uniform(size=(prediction_sets.shape[0], 1))
-    # remove last label from comparison between uniform number and V
-    vs_less_than_us = np.less_equal(vs - us, EPSILON)
-    np.put_along_axis(
-        prediction_sets,
-        y_pred_index_last,
-        vs_less_than_us[:, np.newaxis, :],
-        axis=1
-    )
-    return prediction_sets
